fix: organization card overflow issue for small screen
