+++ apps/gauzy/src/app/pages/organizations/edit-organization/edit-organization-settings/edit-organization-settings.component.html
@@ -10,171 +10,97 @@
 			<nb-icon icon="arrow-back-outline"></nb-icon>
 			<span>{{ 'BUTTONS.BACK' | translate }}</span>
 		</div>
-		<div class="content">
-			<div class="employee-container">
-				<div class="employee-photo">
-					<ngx-image-uploader
-						(changeHoverState)="hoverState = $event"
-						(uploadedImageUrl)="updateImageUrl($event)"
-						(uploadImageError)="handleImageUploadError($event)"
-					></ngx-image-uploader>
-					<svg
-						xmlns="http://www.w3.org/2000/svg"
-						xmlns:xlink="http://www.w3.org/1999/xlink"
-						width="68"
-						height="68"
-						viewBox="0 0 68 68"
-						[ngStyle]="{ opacity: hoverState ? '1' : '0.3' }"
-					>
-						<defs>
-							<path
-								id="a"
-								d="M28.667 31.333a2 2 0 1 0-.002-4.001 2 2 0 0 0 .002 4.001m13.333 12H26.748l9.34-7.793c.328-.279.923-.277 1.244-.001l6.001 5.12V42c0 .736-.597 1.333-1.333 1.333M26 24.667h16c.736 0 1.333.597 1.333 1.333v11.152l-4.27-3.643c-1.32-1.122-3.386-1.122-4.694-.008l-9.702 8.096V26c0-.736.597-1.333 1.333-1.333M42 22H26c-2.205 0-4 1.795-4 4v16c0 2.205 1.795 4 4 4h16c2.205 0 4-1.795 4-4V26c0-2.205-1.795-4-4-4"
-							/>
-						</defs>
-						<g fill="none" fill-rule="evenodd">
-							<circle
-								cx="34"
-								cy="34"
-								r="34"
-								fill="#0091FF"
-								opacity=".3"
-							/>
-							<circle
-								cx="34"
-								cy="34"
-								r="26"
-								fill="#0091FF"
-								opacity=".9"
-							/>
-							<use
-								fill="#FFF"
-								fill-rule="nonzero"
-								xlink:href="#a"
-							/>
-						</g>
-					</svg>
-					<div
-						class="image-overlay"
-						[ngStyle]="{ opacity: hoverState ? '0.2' : '0' }"
-					></div>
-
-					<img
-						*ngIf="imageUrl"
-						[src]="imageUrl"
-						alt="Organization Photo"
-						(mouseenter)="hoverState = true"
-						(mouseleave)="hoverState = false"
-					/>
-				</div>
-				<h6 class="text-center mt-2">
-					{{ employeesCount }}
-					{{ 'ORGANIZATIONS_PAGE.EMPLOYEES' | translate }}
-				</h6>
-			</div>
-
-			<div *ngIf="organization" class="org-settings">
-				<nb-tabset
-					(changeTab)="
-						tabChange($event);
-						clients.loadClients();
-						clients.showAddCard = false;
-						projects.loadProjects();
-						projects.showAddCard = false
-					"
-					fullWidth
+		<div *ngIf="organization" class="org-settings">
+			<nb-tabset
+				(changeTab)="
+					tabChange($event);
+					clients.loadClients();
+					clients.showAddCard = false;
+					projects.loadProjects();
+					projects.showAddCard = false
+				"
+				fullWidth
+			>
+				<nb-tab
+					[active]="activeTabName === 'main'"
+					tabTitle="{{ 'ORGANIZATIONS_PAGE.MAIN' | translate }}"
+					tabIcon="person-outline"
+					responsive
 				>
-					<nb-tab
-						[active]="activeTabName === 'main'"
-						tabTitle="{{ 'ORGANIZATIONS_PAGE.MAIN' | translate }}"
-					>
-						<ga-edit-org-main
-							#main
-							[organization]="organization"
-							[countries]="countries"
-						></ga-edit-org-main>
+					<ga-edit-org-main
+						[organization]="organization"
+						[countries]="countries"
+					></ga-edit-org-main>
+				</nb-tab>
 
-						<div class="actions">
-							<button
-								[disabled]="main.form.invalid"
-								(click)="updateOrganizationSettings()"
-								nbButton
-								status="success"
-							>
-								Save
-							</button>
-						</div>
-					</nb-tab>
+				<nb-tab
+					[active]="activeTabName === 'departments'"
+					tabTitle="{{
+						'ORGANIZATIONS_PAGE.DEPARTMENTS' | translate
+					}}"
+					tabIcon="briefcase-outline"
+					responsive
+				>
+					<ga-edit-org-departments
+						[organizationId]="organization?.id"
+					>
+					</ga-edit-org-departments>
+				</nb-tab>
 
-					<nb-tab
-						[active]="activeTabName === 'departments'"
-						tabTitle="{{
-							'ORGANIZATIONS_PAGE.DEPARTMENTS' | translate
-						}}"
-					>
-						<ga-edit-org-departments
-							[organizationId]="organization?.id"
-						>
-						</ga-edit-org-departments>
-					</nb-tab>
+				<nb-tab
+					[active]="activeTabName === 'clients'"
+					tabTitle="{{ 'ORGANIZATIONS_PAGE.CLIENTS' | translate }}"
+					tabIcon="book-open-outline"
+					responsive
+				>
+					<ga-edit-org-clients
+						#clients
+						[organizationId]="organization?.id"
+					></ga-edit-org-clients>
+				</nb-tab>
 
-					<nb-tab
-						[active]="activeTabName === 'clients'"
-						tabTitle="{{
-							'ORGANIZATIONS_PAGE.CLIENTS' | translate
-						}}"
-					>
-						<ga-edit-org-clients
-							#clients
-							[organizationId]="organization?.id"
-						></ga-edit-org-clients>
-					</nb-tab>
+				<nb-tab
+					[active]="activeTabName === 'positions'"
+					tabTitle="{{ 'ORGANIZATIONS_PAGE.POSITIONS' | translate }}"
+					tabIcon="award-outline"
+					responsive
+				>
+					<ga-edit-org-positions
+						[organizationId]="organization?.id"
+					></ga-edit-org-positions>
+				</nb-tab>
 
-					<nb-tab
-						[active]="activeTabName === 'positions'"
-						tabTitle="{{
-							'ORGANIZATIONS_PAGE.POSITIONS' | translate
-						}}"
-					>
-						<ga-edit-org-positions
-							[organizationId]="organization?.id"
-						></ga-edit-org-positions>
-					</nb-tab>
-
-					<nb-tab
-						[active]="activeTabName === 'vendors'"
-						tabTitle="{{
-							'ORGANIZATIONS_PAGE.VENDORS' | translate
-						}}"
-					>
-						<ga-edit-org-vendors
-							[organizationId]="organization?.id"
-						>
-						</ga-edit-org-vendors>
-					</nb-tab>
-					<nb-tab
-						[active]="activeTabName === 'projects'"
-						tabTitle="{{
-							'ORGANIZATIONS_PAGE.PROJECTS' | translate
-						}}"
-					>
-						<ga-edit-org-projects
-							#projects
-							[organizationId]="organization?.id"
-						></ga-edit-org-projects>
-					</nb-tab>
-					<nb-tab
-						[active]="activeTabName === 'teams'"
-						tabTitle="{{
-							'ORGANIZATIONS_PAGE.EDIT.TEAMS' | translate
-						}}"
-					>
-						<ga-edit-org-teams>
-							#teams [organizationId]="organization?.id"
-						</ga-edit-org-teams>
-					</nb-tab>
-				</nb-tabset>
-			</div>
+				<nb-tab
+					[active]="activeTabName === 'vendors'"
+					tabTitle="{{ 'ORGANIZATIONS_PAGE.VENDORS' | translate }}"
+					tabIcon="car-outline"
+					responsive
+				>
+					<ga-edit-org-vendors [organizationId]="organization?.id">
+					</ga-edit-org-vendors>
+				</nb-tab>
+				<nb-tab
+					[active]="activeTabName === 'projects'"
+					tabTitle="{{ 'ORGANIZATIONS_PAGE.PROJECTS' | translate }}"
+					tabIcon="book-outline"
+					responsive
+				>
+					<ga-edit-org-projects
+						#projects
+						[organizationId]="organization?.id"
+					></ga-edit-org-projects>
+				</nb-tab>
+				<nb-tab
+					[active]="activeTabName === 'teams'"
+					tabTitle="{{ 'ORGANIZATIONS_PAGE.EDIT.TEAMS' | translate }}"
+					tabIcon="people-outline"
+					responsive
+				>
+					<ga-edit-org-teams>
+						#teams [organizationId]="organization?.id"
+					</ga-edit-org-teams>
+				</nb-tab>
+			</nb-tabset>
 		</div>
 	</nb-card-body>
 	<nb-card-footer> </nb-card-footer>
