+++ apps/gauzy/src/app/pages/organizations/edit-organization/edit-organization-settings/edit-organization-settings.component.scss
@@ -10,3 +10,11 @@ nb-tab,
 nb-tabset {
 	height: 100%;
 }
+
+nb-card-body {
+	overflow-y: hidden; //prevent extra scroll bar
+}
+
+:host ::ng-deep .tab-link {
+	padding: 1rem 5px !important;
+}
